pyproject.toml
[project]
name = "personal-cloud-service"
version = "0.1"
dependencies = ["fastapi", "uvicorn", "pytest", "httpx", "python-multipart", "psutil", "python-jose[cryptography]", "passlib[bcrypt]", "python-dotenv", "google-auth", "pyyaml", "requests", "pytest-asyncio", ]